[synaptics-led] v3.6 (and PKGBUILD v3)
There's a bit of cleanup to the PKGBUILD. The current state (which isn't
much different than before) is that this package builds against the
currently running kernel. Unfortunately, this leaves the possibility
that if you have an older kernel, you will be installing psmouse.ko
built from a newer source.